requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.17.0
six==1.15.0
Werkzeug==1.0.1
uwsgi
